feat: remake front for friend page (css, html) and fix socket for addfriend, removefriend, ...
