+++ package.json
@@ -16,9 +16,9 @@
     "dotenv": "^16.4.5",
     "emoji-mart": "^5.6.0",
     "font-awesome": "^4.7.0",
+    "framer-motion": "^11.11.17",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-icons": "^5.3.0",
     "react-router-dom": "^6.27.0",
     "socket.io-client": "^4.8.1",
     "uuid": "^11.0.3"
