+++ src/Components/Navbar/Navbar.css
@@ -208,11 +208,6 @@
   box-sizing: border-box;
   overflow-y: auto;
   z-index: 1000;
-  transition: transform 0.3s ease-out;
-}
-
-.navbar-no-side-popup-closed {
-  transform: translateX(-100%);
 }
 
 .navbar-add-friend-back-btn {
